SyncStatusFormatter: extract resource formatters.
Extract resource formatters to separated files
