+++ detox/src/client/actions/formatters/sync-resources/DelayedPerformSelectorFormatter.js
@@ -1,0 +1,7 @@
+const { makeResourceTitle } = require('./utils');
+
+module.exports = function(properties) {
+  return makeResourceTitle(
+    `There are ${properties.pending_selectors} pending delayed selectors to be performed.`
+  );
+};
