+++ detox/src/client/actions/formatters/sync-resources/DispatchQueueFormatter.js
@@ -1,0 +1,7 @@
+const { makeResourceTitle } = require('./utils');
+
+module.exports = function(properties) {
+  return makeResourceTitle(
+    `There are ${properties.works_count} work items pending on the dispatch queue: "${properties.queue}".`
+  );
+};
