+++ detox/src/client/actions/formatters/sync-resources/RunLoopFormatter.js
@@ -1,0 +1,5 @@
+const { makeResourceTitle } = require('./utils');
+
+module.exports = function(properties) {
+  return makeResourceTitle(`Run loop "${properties.name}" is awake.`);
+};
