+++ detox/src/client/actions/formatters/sync-resources/UnknownResourceFormatter.js
@@ -1,0 +1,5 @@
+const { makeResourceTitle } = require('./utils');
+
+module.exports = function(properties) {
+  return makeResourceTitle(`Resource "${properties.identifier}" is busy.`);
+};
