+++ detox/src/client/actions/formatters/sync-resources/utils.js
@@ -1,0 +1,17 @@
+function makeResourceTitle(string) {
+  return `• ${string}`;
+}
+
+function makeResourceSubTitle(string) {
+  return `  - ${string}`;
+}
+
+function makeResourceSubSubTitle(string) {
+  return `    + ${string}`;
+}
+
+module.exports = {
+  makeResourceTitle,
+  makeResourceSubTitle,
+  makeResourceSubSubTitle,
+};
